oh my fucking god, so many changes; Its 00:20 am and I cant think straight, any code written from now on can be considered trash.
Some changes:
no more logPort, now there is the chatPort. Ill use it to send the chats to the client, the log is now gonna send the messages normally.

Now the client sends the version, username and password all at once, something like this: "1.0.0;username;password"

Thats it, I think